Eliminate redundant casts.

+++ Tests/Caravela.Framework.Tests.InternalPipeline/Tests/Templating/Syntax/Lambdas/CompileTimeArgumentInSimpleLambda.t.cs
@@ -2,5 +2,5 @@ int Method(int a, int b)
 {
     global::System.Action<global::System.Object?> action = a_1 => global::System.Console.WriteLine(a_1?.ToString());
     action(1);
-    return (int)this.Method(a, b);
+    return this.Method(a, b);
 }
