+++ Tests/Caravela.Framework.Tests.InternalPipeline/Tests/Templating/Syntax/Tuples/RuntimeTuples2.t.cs
@@ -2,5 +2,5 @@ int Method(int a)
 {
     var t = (1, 2, 3);
     global::System.Console.WriteLine(t.Item3);
-    return (int)this.Method(a);
+    return this.Method(a);
 }
